Cleanup: move autosuggest-related code into search module
